dash/requirements.txt
pandas
plotly.express
dash
dash_core_components
dash_html_components
